+++ util.h
@@ -2,27 +2,28 @@
 #define _SP_FS_UTIL_H
 
 #include <linux/buffer_head.h>
+#include <linux/types.h>
 
 extern size_t
 spfs_sb_remaining(struct buffer_head *bh, size_t pos);
 
 extern bool
-spfs_sb_read_u16(struct buffer_head *bh, size_t *pos, unsigned short *out);
+spfs_sb_read_u16(struct buffer_head *bh, size_t *pos, uint16_t *out);
 
 extern bool
-spfs_sb_write_u16(struct buffer_head *bh, size_t *pos, unsigned short val);
+spfs_sb_write_u16(struct buffer_head *bh, size_t *pos, uint16_t val);
 
 extern bool
-spfs_sb_read_u32(struct buffer_head *bh, size_t *pos, unsigned long *out);
+spfs_sb_read_u32(struct buffer_head *bh, size_t *pos, uint32_t *out);
 
 extern bool
-spfs_sb_write_u32(struct buffer_head *bh, size_t *pos, unsigned long val);
+spfs_sb_write_u32(struct buffer_head *bh, size_t *pos, uint32_t val);
 
 extern bool
-spfs_sb_read_u64(struct buffer_head *bh, size_t *pos, unsigned long long *out);
+spfs_sb_read_u64(struct buffer_head *bh, size_t *pos, uint64_t *out);
 
 extern bool
-spfs_sb_write_u64(struct buffer_head *bh, size_t *pos, unsigned long long val);
+spfs_sb_write_u64(struct buffer_head *bh, size_t *pos, uint64_t val);
 
 extern bool
 spfs_sb_read_str(struct buffer_head *bh, size_t *pos, char *str, size_t len);
